Refactor irNull and IRFields a bunch.
* convert is now "fromStorage"
* convertFromInput is now "fromInput"

Instead of subclasses overriding these, IRField implements the outer
wrapper (which handles irNull consistantly across all types, and maybe
other things in the future).

The subclasses now implement the underscore-prefixed function:

* _fromStorage
* _fromInput
* _toStorage

Move most all of the irNull work out of IndexedRedisModel and associated
Helper classes, and move it into IRField directly.

There still is some additional work to be done, but it is mostly
functional, and cleans things up a lot.