Make better the way REDIS_CONNECTION_PARAMS work
  You no longer need to define this on every model. If
  REDIS_CONNECTION_PARAMS is not defined on a model or is empty,
  "default" connection params will be used.

  set/get the default connection params via
  setDefaultRedisConnectionParams / getDefaultRedisConnectionParams

Also, change the name of "connect" to "connectAlt", to better represent
what that method does (returns a model with an alternate connection
defined).

Also, update connectAlt to use a unique class name for the copy, to
better support multiple copies.

Also, update copyModel to use a deepcopy of the model dict when copying,
to ensure things aren't linked that shouldn't be linked.